refactor code
- refactor type in different files
