+++ components/animalcard/Animalcard.tsx
@@ -1,7 +1,14 @@
 import styles from "./Animalcard.module.css";
-import { Animal } from "../../utils/type";
 
-function AnimalCard({ name, ImgSrc, description, audioFile }: Animal) {
+export type AnimalCardProps = {
+  _id: string;
+  name: string;
+  ImgSrc: string;
+  description: string;
+  audioFile: string;
+};
+
+function AnimalCard({ name, ImgSrc, description, audioFile }: AnimalCardProps) {
   return (
     <div className={styles.background}>
       <h1 className={styles.headline}>{name}</h1>
