+++ utils/type.ts
@@ -5,13 +5,3 @@ export type Animal = {
   description: string;
   audioFile: string;
 };
-
-export type QuizProps = {
-  _id: string;
-  question: string;
-  answerOptions: Answer[];
-};
-export type Answer = {
-  answerText: string;
-  isCorrect: boolean;
-};
